ui: fix spritemap generation for assist sessions
