+++ frontend/app/player/web/MessageManager.ts
@@ -201,7 +201,15 @@ export default class MessageManager {
     }
 
     Object.values(this.tabs).forEach((tab) => tab.onFileReadSuccess?.());
+
+    this.updateSpriteMap();
   };
+
+  public updateSpriteMap = () => {
+    if (this.spriteMapSvg) {
+      this.injectSpriteMap(this.spriteMapSvg);
+    }
+  }
 
   public onFileReadFailed = (...e: any[]) => {
     logger.error(e);
@@ -337,9 +345,38 @@ export default class MessageManager {
     this.state.update({ tabChangeEvents: this.tabChangeEvents });
   }
 
+  spriteMapSvg: SVGElement | null = null;
+  potentialSpriteMap: Record<string, any> = {};
+  domParser: DOMParser | null = null;
+  createSpriteMap = () => {
+    if (!this.spriteMapSvg) {
+      this.domParser = new DOMParser();
+      this.spriteMapSvg = document.createElementNS(
+        'http://www.w3.org/2000/svg',
+        'svg',
+      );
+      this.spriteMapSvg.setAttribute('style', 'display: none;');
+      this.spriteMapSvg.setAttribute('id', 'reconstructed-sprite');
+    }
+  };
+
   distributeMessage = (msg: Message & { tabId: string }): void => {
     // @ts-ignore placeholder msg for timestamps
     if (msg.tp === 9999) return;
+    if (msg.tp === MType.SetNodeAttribute) {
+      if (msg.value.includes('_$OPENREPLAY_SPRITE$_')) {
+        this.createSpriteMap();
+        if (!this.domParser) {
+          return console.error('DOM parser is not initialized?');
+        }
+        handleSprites(
+          this.potentialSpriteMap,
+          this.domParser,
+          msg,
+          this.spriteMapSvg!,
+        );
+      }
+    }
     if (!this.tabs[msg.tabId]) {
       this.tabsAmount++;
       this.state.update({
@@ -453,4 +490,37 @@ function mapTabs(tabs: Record<string, TabSessionManager>) {
   });
 
   return tabMap;
+}
+
+function handleSprites(
+  potentialSpriteMap: Record<string, any>,
+  parser: DOMParser,
+  msg: Record<string, any>,
+  spriteMapSvg: SVGElement,
+) {
+  const [_, svgData] = msg.value.split('_$OPENREPLAY_SPRITE$_');
+  const potentialSprite = potentialSpriteMap[svgData];
+  if (potentialSprite) {
+    msg.value = potentialSprite;
+  } else {
+    const svgDoc = parser.parseFromString(svgData, 'image/svg+xml');
+    const originalSvg = svgDoc.querySelector('svg');
+    if (originalSvg) {
+      const symbol = document.createElementNS(
+        'http://www.w3.org/2000/svg',
+        'symbol',
+      );
+      const symbolId = `symbol-${msg.id || `ind-${msg.time}`}`; // Generate an ID if missing
+      symbol.setAttribute('id', symbolId);
+      symbol.setAttribute(
+        'viewBox',
+        originalSvg.getAttribute('viewBox') || '0 0 24 24',
+      );
+      symbol.innerHTML = originalSvg.innerHTML;
+
+      spriteMapSvg.appendChild(symbol);
+      msg.value = `#${symbolId}`;
+      potentialSpriteMap[svgData] = `#${symbolId}`;
+    }
+  }
 }
